files/requirements.txt
flask
torch
torchvision
transformers
huggingface-hub
pillow
requests
timm
